refactor: changed to use tsconfig paths compiler options
