+++ src/algorithms-core/arrays_common.ts
